Revert "Resolve the Ingestion rate limit exceeded (limit: 10485760 bytes/sec) while attempting to ingest"
This reverts commit 0c61958522bd6af5f425c9db41fa60fc679dffe5.

This fix doesn't work as expected. The Observatorium operator overwrites
the configMap settings. Would need to reapply the kustomization after
operator overwrote the configMap which is ugly.

The distributor rate limit configuration doesn't seem to be supported
by the Observatorium operator at this time. Instead, let's scale the
number of distributor and ingestor replicas up, so that we can process
the incoming logs without hitting the rate limiting thresholds.